move can_id pack details into ocan.
